move handshake fn to handshake file
